+++ backend/users/views.py
@@ -1,0 +1,25 @@
+from django.contrib.auth import get_user_model
+from rest_framework import generics, permissions
+
+from .serializers import RegisterSerializer, UserSerializer
+
+User = get_user_model()
+
+#
+class RegisterView(generics.CreateAPIView):
+    """
+    Vue pour l'inscription d'un nouvel utilisateur.
+    """
+    queryset = User.objects.all()
+    permission_classes = [permissions.AllowAny]
+    serializer_class = RegisterSerializer
+
+class MeView(generics.RetrieveUpdateAPIView):
+    """
+    Vue pour récupérer les informations de l'utilisateur connecté.
+    """
+    serializer_class = UserSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get_object(self):
+        return self.request.user